svn merge --revision 24924:25057 -x --ignore-eol-style --non-interactive https://svn.tribler.org/abc/branches/release-5.5.x


git-svn-id: https://svn.tribler.org/abc/branches/mainbranch@25058 001aeff7-3401-0410-a489-f7902fc005dd

+++ Tribler/Core/APIImplementation/UserCallbackHandler.py
@@ -125,19 +125,20 @@ class UserCallbackHandler:
             if len(contentdests) > 1:
                 def remove_if_empty(basedir):
                     #first try to remove sub-dirs
-                    files = os.listdir(basedir)
-                    for filename in files:
-                        absfilename = os.path.join(basedir, filename)
-                        if os.path.isdir(absfilename) and absfilename in contentdirs:
-                            remove_if_empty(absfilename)
-                    
-                    #see if we are empty
-                    files = os.listdir(basedir)
-                    #ignore thumbs.db files
-                    files = [file for file in files if not file.lower().endswith('thumbs.db')]
-                    
-                    if len(files) == 0:
-                        os.rmdir(basedir)
+                    if os.path.isdir(basedir):
+                        files = os.listdir(basedir)
+                        for filename in files:
+                            absfilename = os.path.join(basedir, filename)
+                            if os.path.isdir(absfilename) and absfilename in contentdirs:
+                                remove_if_empty(absfilename)
+                        
+                        #see if we are empty
+                        files = os.listdir(basedir)
+                        #ignore thumbs.db files
+                        files = [file for file in files if not file.lower().endswith('thumbs.db')]
+                        
+                        if len(files) == 0:
+                            os.rmdir(basedir)
                 
                 basedir = os.path.commonprefix(contentdests)
                 remove_if_empty(basedir)
